Expand Capabilities to support more workflows
MagicValueSorcerer is now much more generic and supports one sorcerer supports generating multiple files.
Defaults for the included Sorcerers have been moved to the DefaultMagicValueSorcerer.
SimpleMagicValueSorcerer has also been created to give a more simple API that is similiar to the default API for clients to use when they don't need all the customizability that MagicValueSorcerer now provides.
ClassBuilder now supports more static fields, blank lines, and has a more robust CleanName function.
Moved Settings from Preferences to Project because these affect the project and should be shared across a team and are not user preferences.
Also did some cleanup of the settings window to wrap the tab rows when there is more than 5 tabs to display instead of making them impossibly small with 10+ tabs to show.
Settings key names also changed slightly.
MagicValueSorcerers are no longer registered with an attribute and we instead rely on Unity to find all instances of MagicValueSorcerer.